Add a package avro under implementation and move the corresponding classes
