+++ lang/java/reef-tang/tang/src/test/java/org/apache/reef/tang/test/RoundTripTest.java
@@ -21,7 +21,7 @@ package org.apache.reef.tang.test;
 import org.apache.reef.tang.ClassHierarchy;
 import org.apache.reef.tang.Configuration;
 import org.apache.reef.tang.Tang;
-import org.apache.reef.tang.formats.AvroClassHierarchySerializer;
+import org.apache.reef.tang.implementation.avro.AvroClassHierarchySerializer;
 import org.apache.reef.tang.formats.ClassHierarchySerializer;
 import org.junit.Assert;
 import org.junit.Test;
